+++ copulas/multivariate/tree.py
@@ -1,4 +1,6 @@
+import json
 import logging
+from enum import Enum
 
 import numpy as np
 import scipy
@@ -9,12 +11,56 @@ from copulas.bivariate.base import Bivariate
 LOGGER = logging.getLogger(__name__)
 
 
-class Tree(object):
-    """Helper class to instantiate a single tree in the vine model
-    """
+class TreeTypes(Enum):
+    CENTER = 0
+    DIRECT = 1
+    REGULAR = 2
 
-    def __init__(self, index, n_nodes, tau_matrix, previous_tree):
-        """Initialize tree object
+
+class Tree(object):
+    """Helper class to instantiate a single tree in the vine model."""
+
+    tree_type = None
+    _subclasses = []
+
+    @classmethod
+    def _get_subclasses(cls):
+        subclasses = []
+        for subclass in cls.__subclasses__():
+            subclasses.append(subclass)
+            subclasses.extend(subclass._get_subclasses())
+
+        return subclasses
+
+    @classmethod
+    def subclasses(cls):
+        if not cls._subclasses:
+            cls._subclasses = cls._get_subclasses()
+
+        return cls._subclasses
+
+    def __new__(cls, tree_type=None):
+        if not isinstance(tree_type, TreeTypes):
+            if (isinstance(tree_type, str) and tree_type.upper() in TreeTypes.__members__):
+                tree_type = TreeTypes[tree_type.upper()]
+            else:
+                raise ValueError('Invalid tree type {}'.format(tree_type))
+
+        for subclass in cls.subclasses():
+            if subclass.tree_type is tree_type:
+                return super(Tree, cls).__new__(subclass)
+
+    def __init__(self, tree_type=None):
+        """
+        Creates a new instance of any of their subclasses.
+
+        Args:
+            tree_type: `TreeType` or `str` to be compared against  TreeType.
+        """
+        pass
+
+    def fit(self, index, n_nodes, tau_matrix, previous_tree, edges=None):
+        """Fits tree object.
 
         Args:
             :param index: index of the tree
@@ -26,19 +72,23 @@ class Tree(object):
             :type tau_matrix: np.ndarray of size n_nodes*n_nodes
         """
         self.level = index + 1
-        self.previous_tree = previous_tree
         self.n_nodes = n_nodes
-        self.edges = []
         self.tau_matrix = tau_matrix
-        if self.level == 1:
-            self.u_matrix = previous_tree
-            self._build_first_tree()
-        else:
-            self._build_kth_tree()
-        self.prepare_next_tree()
+        self.previous_tree = previous_tree
+        self.edges = edges or []
+
+        if not self.edges:
+            if self.level == 1:
+                self.u_matrix = previous_tree
+                self._build_first_tree()
+
+            else:
+                self._build_kth_tree()
+
+            self.prepare_next_tree()
 
     def _check_contraint(self, edge1, edge2):
-        """Check if two edges satisfy vine constraint
+        """Check if two edges satisfy vine constraint.
 
         Args:
             :param edge1: edge object representing edge1
@@ -55,17 +105,16 @@ class Tree(object):
         return len(full_node) == (self.level + 1)
 
     def _get_constraints(self):
-        """Get neighboring edges for each edge in the edges"""
+        """Get neighboring edges for each edge in the edges."""
         num_edges = len(self.edges)
         for k in range(num_edges):
             for i in range(num_edges):
-                # add to constriants if i shared an edge with k
-                if k != i:
-                    if self.edges[k].is_adjacent(self.edges[i]):
-                        self.edges[k].neighbors.append(i)
+                # add to constraints if i shared an edge with k
+                if k != i and self.edges[k].is_adjacent(self.edges[i]):
+                    self.edges[k].neighbors.append(i)
 
     def _sort_tau_by_y(self, y):
-        """Sort tau matrix by dependece with variable y
+        """Sort tau matrix by dependece with variable y.
 
         Args:
             :param y: index of variable of intrest
@@ -74,16 +123,18 @@ class Tree(object):
         # first column is the variable of interest
         tau_y = self.tau_matrix[:, y]
         tau_y[y] = np.NaN
+
         temp = np.empty([self.n_nodes, 3])
         temp[:, 0] = np.arange(self.n_nodes)
         temp[:, 1] = tau_y
         temp[:, 2] = abs(tau_y)
         temp[np.isnan(temp)] = -10
         tau_sorted = temp[temp[:, 2].argsort()[::-1]]
+
         return tau_sorted
 
     def get_tau_matrix(self):
-        """Get tau matrix for adjacent pairs
+        """Get tau matrix for adjacent pairs.
 
         Returns:
             :param tau: tau matrix for the current tree
@@ -91,20 +142,24 @@ class Tree(object):
         """
         num_edges = len(self.edges)
         tau = np.empty([num_edges, num_edges])
+
         for i in range(num_edges):
             edge = self.edges[i]
             for j in edge.neighbors:
                 if self.level == 1:
                     left_u = self.u_matrix[:, edge.L]
                     right_u = self.u_matrix[:, edge.R]
+
                 else:
                     left_parent, right_parent = edge.parents
                     left_u, right_u = Edge.get_conditional_uni(left_parent, right_parent)
+
                 tau[i, j], pvalue = scipy.stats.kendalltau(left_u, right_u)
+
         return tau
 
     def get_adjacent_matrix(self):
-        """ Get adjacency matrix
+        """Get adjacency matrix.
 
         Returns:
             :param adj: adjacency matrix
@@ -113,32 +168,37 @@ class Tree(object):
         edges = self.edges
         num_edges = len(edges) + 1
         adj = np.zeros([num_edges, num_edges])
+
         for k in range(num_edges - 1):
             adj[edges[k].L, edges[k].R] = 1
             adj[edges[k].R, edges[k].L] = 1
+
         return adj
 
     def prepare_next_tree(self):
-        """Prepare conditional U matrix for next tree
-        """
-        num_edges = len(self.edges)
-        for k in range(num_edges):
-            edge = self.edges[k]
+        """Prepare conditional U matrix for next tree."""
+        for edge in self.edges:
             copula_theta = edge.theta
+
             if self.level == 1:
                 left_u = self.u_matrix[:, edge.L]
                 right_u = self.u_matrix[:, edge.R]
+
             else:
                 left_parent, right_parent = edge.parents
                 left_u, right_u = Edge.get_conditional_uni(left_parent, right_parent)
+
             # compute conditional cdfs C(i|j) = dC(i,j)/duj and dC(i,j)/du
             left_u = [x for x in left_u if x is not None]
             right_u = [x for x in right_u if x is not None]
-            cop = Bivariate(edge.name)
-            cop.fit(left_u, right_u)
-            derivative = cop.get_h_function()
-            left_given_right = derivative(left_u, right_u, copula_theta)
-            right_given_left = derivative(right_u, left_u, copula_theta)
+            X_left_right = np.array([[x, y] for x, y in zip(left_u, right_u)])
+            X_right_left = np.array([[x, y] for x, y in zip(right_u, left_u)])
+
+            copula = Bivariate(edge.name)
+            copula.fit(X_left_right)
+            left_given_right = copula.partial_derivative(X_left_right, copula_theta)
+            right_given_left = copula.partial_derivative(X_right_left, copula_theta)
+
             # correction of 0 or 1
             left_given_right[left_given_right == 0] = EPSILON
             right_given_left[right_given_left == 0] = EPSILON
@@ -147,7 +207,7 @@ class Tree(object):
             edge.U = [left_given_right, right_given_left]
 
     def get_likelihood(self, uni_matrix):
-        """Compute likelihood of the tree given an U matrix
+        """Compute likelihood of the tree given an U matrix.
 
         Args:
             :param uni_matrix: univariate matrix to evaluate likelihood on
@@ -163,12 +223,14 @@ class Tree(object):
         num_edge = len(self.edges)
         values = np.zeros([1, num_edge])
         new_uni_matrix = np.empty([uni_dim, uni_dim])
+
         for i in range(num_edge):
             edge = self.edges[i]
             value, left_u, right_u = edge.get_likelihood(uni_matrix)
             new_uni_matrix[edge.L, edge.R] = left_u
             new_uni_matrix[edge.R, edge.L] = right_u
             values[0, i] = np.log(value)
+
         return np.sum(values), new_uni_matrix
 
     def __str__(self):
@@ -176,26 +238,57 @@ class Tree(object):
         return '\n'.join([template.format(edge.L, edge.R, edge.D, edge.name, edge.theta)
                           for edge in self.edges])
 
+    def to_dict(self):
+        return {
+            'level': self.level - 1,
+            'n_nodes': self.n_nodes,
+            'tau_matrix': self.tau_matrix.tolist(),
+            'previous_tree': self.previous_tree.tolist(),
+            'edges': [edge.to_dict() for edge in self.edges],
+        }
+
+    @classmethod
+    def from_dict(cls, tree_dict):
+        """Create a new instance from a dictionary."""
+        return cls(**tree_dict)
+
+    @classmethod
+    def load(cls, tree_path):
+        """Create a new instance from a file."""
+        with open(tree_path) as f:
+            tree_dict = json.load(f)
+
+        return cls.from_dict(tree_dict)
+
+    def save(self, filename):
+        """Save the internal state of a copula in the specified filename."""
+        content = self.to_dict()
+        with open(filename, 'w') as f:
+            json.dump(content, f)
+
 
 class CenterTree(Tree):
     """Helper Class for instantiate a Center Vine"""
 
+    tree_type = TreeTypes.CENTER
+
     def _build_first_tree(self):
-        """build first level tree"""
+        """Build first level tree."""
         tau_sorted = self._sort_tau_by_y(0)
         for itr in range(self.n_nodes - 1):
             ind = int(tau_sorted[itr, 0])
-            name, theta = Bivariate.select_copula(self.u_matrix[:, 0],
-                                                  self.u_matrix[:, ind])
+            name, theta = Bivariate.select_copula(self.u_matrix[:, (0, ind)])
+
             new_edge = Edge(0, ind, name, theta)
             new_edge.tau = self.tau_matrix[0, ind]
             self.edges.append(new_edge)
 
     def _build_kth_tree(self):
-        """build k-th level tree"""
+        """Build k-th level tree."""
         anchor = self.get_anchor()
         aux_sorted = self._sort_tau_by_y(anchor)
         edges = self.previous_tree.edges
+
         for itr in range(self.n_nodes - 1):
             right = int(aux_sorted[itr, 0])
             left_parent, right_parent = Edge.sort_edge([edges[anchor], edges[right]])
@@ -204,7 +297,7 @@ class CenterTree(Tree):
             self.edges.append(new_edge)
 
     def get_anchor(self):
-        """find anchor variable with highest sum of dependence with the rest"""
+        """Find anchor variable with highest sum of dependence with the rest."""
         temp = np.empty([self.n_nodes, 2])
         temp[:, 0] = np.arange(self.n_nodes, dtype=int)
         temp[:, 1] = np.sum(abs(self.tau_matrix), 1)
@@ -213,7 +306,9 @@ class CenterTree(Tree):
 
 
 class DirectTree(Tree):
-    """Helper Class for instantiate a Direct Vine"""
+    """Helper Class for instantiate a Direct Vine."""
+
+    tree_type = TreeTypes.DIRECT
 
     def _build_first_tree(self):
         # find the pair of maximum tau
@@ -223,26 +318,30 @@ class DirectTree(Tree):
         right_ind = tau_sorted[1, 0]
         T1 = np.array([left_ind, 0, right_ind]).astype(int)
         tau_T1 = tau_sorted[:2, 1]
+
         # replace tau matrix of the selected variables as a negative number
         tau_matrix[:, [T1]] = -10
         for k in range(2, self.n_nodes - 1):
-            valL, left = np.max(tau_matrix[T1[0], :]),\
-                np.argmax(tau_matrix[T1[0], :])
-            valR, right = np.max(tau_matrix[T1[-1], :]),\
-                np.argmax(tau_matrix[T1[-1], :])
+            left = np.argmax(tau_matrix[T1[0], :])
+            right = np.argmax(tau_matrix[T1[-1], :])
+            valL = np.max(tau_matrix[T1[0], :])
+            valR = np.max(tau_matrix[T1[-1], :])
+
             if valL > valR:
                 # add nodes to the left
                 T1 = np.append(int(left), T1)
                 tau_T1 = np.append(valL, tau_T1)
                 tau_matrix[:, left] = -10
+
             else:
                 # add node to the right
                 T1 = np.append(T1, int(right))
                 tau_T1 = np.append(tau_T1, valR)
                 tau_matrix[:, right] = -10
+
         for k in range(self.n_nodes - 1):
-            name, theta = Bivariate.select_copula(self.u_matrix[:, T1[k]],
-                                                  self.u_matrix[:, T1[k + 1]])
+            name, theta = Bivariate.select_copula(self.u_matrix[:, (T1[k], T1[k + 1])])
+
             left, right = sorted([T1[k], T1[k + 1]])
             new_edge = Edge(left, right, name, theta)
             new_edge.tau = tau_T1[k]
@@ -258,65 +357,67 @@ class DirectTree(Tree):
 
 
 class RegularTree(Tree):
-    """ Helper class for instantiate Regular Vine"""
+    """Helper class for instantiate Regular Vine."""
+
+    tree_type = TreeTypes.REGULAR
 
     def _build_first_tree(self):
-        """build the first tree with n-1 variable"""
+        """Build the first tree with n-1 variable."""
         # Prim's algorithm
         neg_tau = -1.0 * abs(self.tau_matrix)
-        X = set()
-        X.add(0)
-        itr = 0
+        X = {0}
+
         while len(X) != self.n_nodes:
             adj_set = set()
             for x in X:
                 for k in range(self.n_nodes):
                     if k not in X and k != x:
                         adj_set.add((x, k))
+
             # find edge with maximum
             edge = sorted(adj_set, key=lambda e: neg_tau[e[0]][e[1]])[0]
-            name, theta = Bivariate.select_copula(self.u_matrix[:, edge[0]],
-                                                  self.u_matrix[:, edge[1]])
+            name, theta = Bivariate.select_copula(self.u_matrix[:, (edge[0], edge[1])])
+
             left, right = sorted([edge[0], edge[1]])
             new_edge = Edge(left, right, name, theta)
             new_edge.tau = self.tau_matrix[edge[0], edge[1]]
             self.edges.append(new_edge)
             X.add(edge[1])
-            itr += 1
 
     def _build_kth_tree(self):
-        """build tree for level k
-        """
+        """Build tree for level k."""
         neg_tau = -1.0 * abs(self.tau_matrix)
         edges = self.previous_tree.edges
         visited = set([0])
         unvisited = set(range(self.n_nodes))
-        itr = 0
+
         while len(visited) != self.n_nodes:
             adj_set = set()
             for x in visited:
                 for k in range(self.n_nodes):
-                    if k not in visited and k != x:
-                        # check if (x,k) is a valid edge in the vine
-                        if self._check_contraint(edges[x], edges[k]):
-                            adj_set.add((x, k))
+                    # check if (x,k) is a valid edge in the vine
+                    if k not in visited and k != x and self._check_contraint(edges[x], edges[k]):
+                        adj_set.add((x, k))
+
             # find edge with maximum tau
-            if len(list(adj_set)) == 0:
+            if len(adj_set) == 0:
                 visited.add(list(unvisited)[0])
                 continue
+
             pairs = sorted(adj_set, key=lambda e: neg_tau[e[0]][e[1]])[0]
             left_parent, right_parent = Edge.sort_edge([edges[pairs[0]], edges[pairs[1]]])
+
             new_edge = Edge.get_child_edge(left_parent, right_parent)
             new_edge.tau = self.tau_matrix[pairs[0], pairs[1]]
             self.edges.append(new_edge)
+
             visited.add(pairs[1])
             unvisited.remove(pairs[1])
-            itr += 1
 
 
 class Edge(object):
     def __init__(self, left, right, copula_name, copula_theta):
-        """Initialize an Edge object
+        """Initialize an Edge object.
 
         Args:
             :param left: left_node index (smaller)
@@ -339,7 +440,7 @@ class Edge(object):
 
     @staticmethod
     def _identify_eds_ing(edge1, edge2):
-        """Find nodes connecting adjacent edges
+        """Find nodes connecting adjacent edges.
 
         Args:
             :param edge1: edge object representing edge1
@@ -360,7 +461,7 @@ class Edge(object):
         return left, right, depend_set
 
     def is_adjacent(self, another_edge):
-        """Check if two edges are adjacent
+        """Check if two edges are adjacent.
 
         Args:
             :param another_edge: edge object of another edge
@@ -368,59 +469,111 @@ class Edge(object):
 
         This function will return true if the two edges are adjacent
         """
-        return (self.L == another_edge.L or self.L == another_edge.R or
-                self.R == another_edge.L or self.R == another_edge.R)
+        return (
+            self.L == another_edge.L or
+            self.L == another_edge.R or
+            self.R == another_edge.L or
+            self.R == another_edge.R
+        )
 
     @staticmethod
     def sort_edge(edges):
-        """ sort edge object first by left node indices then right
+        """Sort edge object first by left node indices then right.
 
         Args:
             :param edges: list of edges need to be sorted
             :type edges: list
         """
-        sorted_edges = sorted(edges, key=lambda x: (x.L, x.R))
-        return sorted_edges
+        return sorted(edges, key=lambda x: (x.L, x.R))
 
-    @staticmethod
-    def get_conditional_uni(left_parent, right_parent):
-        """ Identify pair univariate value from parents"""
-        left, right, depend_set =\
-            Edge._identify_eds_ing(left_parent, right_parent)
-        if left_parent.L == left:
-            left_u = left_parent.U[0]
-        else:
-            left_u = left_parent.U[1]
-        if right_parent.L == right:
-            right_u = right_parent.U[0]
-        else:
-            right_u = right_parent.U[1]
+    @classmethod
+    def get_conditional_uni(cls, left_parent, right_parent):
+        """Identify pair univariate value from parents."""
+        left, right, depend_set = cls._identify_eds_ing(left_parent, right_parent)
+
+        left_u = left_parent.U[0] if left_parent.L == left else left_parent.U[1]
+        right_u = right_parent.U[0] if right_parent.L == right else right_parent.U[1]
+
         return left_u, right_u
 
-    def get_child_edge(left_parent, right_parent):
-        """ Construct a child edge from two parent edges """
-        [ed1, ed2, depend_set] =\
-            Edge._identify_eds_ing(left_parent, right_parent)
-        left_u, right_u = Edge.get_conditional_uni(left_parent, right_parent)
-        name, theta = Bivariate.select_copula(left_u, right_u)
+    @classmethod
+    def get_child_edge(cls, left_parent, right_parent):
+        """Construct a child edge from two parent edges."""
+        [ed1, ed2, depend_set] = cls._identify_eds_ing(left_parent, right_parent)
+        left_u, right_u = cls.get_conditional_uni(left_parent, right_parent)
+        X = np.array([[x, y] for x, y in zip(left_u, right_u)])
+        name, theta = Bivariate.select_copula(X)
         new_edge = Edge(ed1, ed2, name, theta)
         new_edge.D = depend_set
         new_edge.parents = [left_parent, right_parent]
         return new_edge
 
     def get_likelihood(self, uni_matrix):
-        """ Compute likelihood given a U matrix """
+        """Compute likelihood given a U matrix."""
         if self.parents is None:
             left_u = uni_matrix[:, self.L]
             right_u = uni_matrix[:, self.R]
+
         else:
             left_ing = list(self.D - self.parents[0].D)[0]
             right_ing = list(self.D - self.parents[1].D)[0]
             left_u = uni_matrix[self.L, left_ing]
             right_u = uni_matrix[self.R, right_ing]
-        cop = Bivariate(self.name)
-        cop.set_params(theta=self.theta)
-        value = np.sum(cop.get_pdf()(left_u, right_u))
-        left_given_right = cop.get_h_function()(left_u, right_u, self.theta)
-        right_given_left = cop.get_h_function()(right_u, left_u, self.theta)
+
+        copula = Bivariate(self.name)
+        copula.theta = self.theta
+
+        X_left_right = np.array([[left_u, right_u]])
+        X_right_left = np.array([[right_u, left_u]])
+
+        value = np.sum(copula.probability_density(X_left_right))
+        left_given_right = copula.partial_derivative(X_left_right)
+        right_given_left = copula.partial_derivative(X_right_left)
+
         return value, left_given_right, right_given_left
+
+    def to_dict(self):
+        parents = None
+        if self.parents:
+            parents = [parent.to_dict() for parent in self.parents]
+
+        neighbors = None
+        if self.neighbors:
+            neighbors = [neighbor.to_dict() for neighbor in self.neighbors]
+
+        return {
+            'L': self.L,
+            'R': self.R,
+            'D': self.D,
+            'parents': parents,
+            'neighbors': neighbors,
+            'name': self.name,
+            'theta': self.theta,
+            'tau': self.tau,
+            'U': self.U,
+            'likelihood': self.likelihood
+        }
+
+    @classmethod
+    def from_dict(cls, edge_dict):
+        instance = cls(edge_dict['L'], edge_dict['R'], edge_dict['name'], edge_dict['theta'])
+        parents = edge_dict['parents']
+        neighbors = edge_dict['neighbors']
+
+        if parents:
+            instance.parents = []
+            for parent in parents:
+                edge = Edge.from_dict(parent)
+                instance.parents.append(edge)
+
+        if neighbors:
+            instance.neighbors = []
+            for neighbor in neighbors:
+                edge = Edge.from_dict(neighbor)
+                instance.neighbors.append(edge)
+
+        regular_attributes = ['D', 'tau', 'U', 'likelihood']
+        for key in regular_attributes:
+            setattr(instance, key, edge_dict[key])
+
+        return instance