+++ copulas/univariate/base.py
@@ -4,22 +4,75 @@ class Univariate(object):
     def __init__(self):
         pass
 
-    def fit(self):
-        """ fits a univariate model and updates parameters """
+    def fit(self, X):
+        """Fits the model.
+
+        Arguments:
+            X: `np.ndarray` of shape (n, 1).
+
+        Returns:
+            None
+        """
         raise NotImplementedError
 
-    def get_pdf(self, value):
-        """ given a value, returns corresponding pdf value """
+    def probability_density(self, X):
+        """Computes probability density.
+
+        Arguments:
+            X: `np.ndarray` of shape (n, 1).
+
+        Returns:
+            np.ndarray
+        """
         raise NotImplementedError
 
-    def get_cdf(self, value):
-        """ given a value returns corresponding cdf value """
+    def pdf(self, X):
+        return self.probability_density(X)
+
+    def cumulative_distribution(self, X):
+        """Computes cumulative density.
+
+        Arguments:
+            X: `np.ndarray` of shape (n, 1).
+
+        Returns:
+            np.ndarray: Cumulative density for X.
+        """
         raise NotImplementedError
 
-    def inverse_cdf(self, value):
-        """ given a cdf value, returns a value in original space """
+    def cdf(self, X):
+        return self.cumulative_distribution(X)
+
+    def percent_point(self, U):
+        """Given a cumulative distribution value, returns a value in original space.
+
+        Arguments:
+            U: `np.ndarray` of shape (n, 1) and values in [0,1]
+
+        Returns:
+            `np.ndarray`: Estimated values in original space.
+        """
         raise NotImplementedError
 
-    def sample(self):
-        """ returns new data point based on model """
+    def ppf(self, U):
+        return self.percent_point(U)
+
+    def sample(self, n_samples=1):
+        """Returns new data point based on model.
+
+        Argument:
+            n_samples: `int`
+
+        Returns:
+            np.ndarray: Generated samples
+        """
+        raise NotImplementedError
+
+    def to_dict(self):
+        """Returns parameters to replicate the distribution."""
+        raise NotImplementedError
+
+    @classmethod
+    def from_dict(cls, param_dict):
+        """Create new instance from dictionary."""
         raise NotImplementedError