+++ tests/copulas/multivariate/test_direct_tree.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 import numpy as np
 import pandas as pd
 
-from copulas.multivariate.tree import DirectTree
+from copulas.multivariate.tree import Tree, TreeTypes
 from copulas.univariate.kde import KDEUnivariate
 
 
@@ -16,9 +16,10 @@ class TestDirectTree(TestCase):
         for col in self.data:
             uni = KDEUnivariate()
             uni.fit(self.data[col])
-            self.u_matrix[:, count] = [uni.get_cdf(x) for x in self.data[col]]
+            self.u_matrix[:, count] = [uni.cumulative_distribution(x) for x in self.data[col]]
             count += 1
-        self.tree = DirectTree(0, 4, self.tau_mat, self.u_matrix)
+        self.tree = Tree(TreeTypes.DIRECT)
+        self.tree.fit(0, 4, self.tau_mat, self.u_matrix)
 
     def test_first_tree(self):
         """ Assert 0 is the center node"""
@@ -30,7 +31,7 @@ class TestDirectTree(TestCase):
 
         value, new_u = self.tree.get_likelihood(uni_matrix)
 
-        expected = -5.4620
+        expected = -0.1207611551427385
         assert abs(value - expected) < 10E-3
 
     def test_get_constraints(self):
@@ -52,12 +53,13 @@ class TestDirectTree(TestCase):
         """ Assert second tree likelihood is correct """
         tau = self.tree.get_tau_matrix()
 
-        second_tree = DirectTree(1, 3, tau, self.tree)
+        second_tree = Tree(TreeTypes.DIRECT)
+        second_tree.fit(1, 3, tau, self.tree)
 
         uni_matrix = np.array([[0.1, 0.2, 0.3, 0.4]])
 
         first_value, new_u = self.tree.get_likelihood(uni_matrix)
         second_value, out_u = second_tree.get_likelihood(new_u)
 
-        expected = 0.7819
+        expected = 0.7184205492690413
         assert abs(second_value - expected) < 10E-3